staticfiles/requirements.txt
dash
pandas
django_plotly_dash
dash-html-components
dash-bootstrap-components
requests
matplotlib
pyyaml
